+++ packages/workbox-expiration/src/Plugin.ts
@@ -6,16 +6,16 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {assert} from 'workbox-core/_private/assert.mjs';
-import {cacheNames} from 'workbox-core/_private/cacheNames.mjs';
-import {getFriendlyURL} from 'workbox-core/_private/getFriendlyURL.mjs';
-import {logger} from 'workbox-core/_private/logger.mjs';
-import {WorkboxError} from 'workbox-core/_private/WorkboxError.mjs';
+import {assert} from 'workbox-core/_private/assert.js';
+import {cacheNames} from 'workbox-core/_private/cacheNames.js';
+import {getFriendlyURL} from 'workbox-core/_private/getFriendlyURL.js';
+import {logger} from 'workbox-core/_private/logger.js';
+import {WorkboxError} from 'workbox-core/_private/WorkboxError.js';
 import {registerQuotaErrorCallback}
-  from 'workbox-core/registerQuotaErrorCallback.mjs';
+  from 'workbox-core/registerQuotaErrorCallback.js';
 
-import {CacheExpiration} from './CacheExpiration.mjs';
-import './_version.mjs';
+import {CacheExpiration} from './CacheExpiration.js';
+import './_version.js';
 
 /**
  * This plugin can be used in the Workbox APIs to regularly enforce a
